refactor: 패키지 구조 변경
- session 패키지 추가
